+++ src/main/java/com/cucumber/utils/context/config/TypeRegistryConfiguration.java
@@ -48,7 +48,7 @@ public class TypeRegistryConfiguration implements TypeRegistryConfigurer {
 
 
     private class ScenarioPropsParameterTransformer implements ParameterByTypeTransformer {
-        ObjectMapper objectMapper = new ObjectMapper()
+        private ObjectMapper objectMapper = new ObjectMapper()
                 .enable(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY)
                 .enable(DeserializationFeature.FAIL_ON_TRAILING_TOKENS);