+++ src/main/java/com/cucumber/utils/context/stepdefs/DateTimeSteps.java
@@ -1,10 +1,10 @@
 package com.cucumber.utils.context.stepdefs;
 
+import com.cucumber.utils.context.utils.ScenarioUtils;
+import com.google.inject.Inject;
 import io.cucumber.guice.ScenarioScoped;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import java.time.Duration;
 import java.time.LocalDate;
@@ -16,8 +16,9 @@ import static org.junit.Assert.assertEquals;
 
 @ScenarioScoped
 public class DateTimeSteps {
-    private Logger log = LogManager.getLogger();
     private DateTimeFormatter formatter;
+    @Inject
+    private ScenarioUtils logger;
 
     @Given("DateTime pattern=\"{}\"")
     public void format(String pattern) {
@@ -26,7 +27,7 @@ public class DateTimeSteps {
 
     @Then("DateTime check period from \"{}\" to \"{}\" is {}year(s)")
     public void compareYears(String date1, String date2, Object y) {
-        log.info("Check period from {} to {}", date1, date2);
+        logger.log("Check period from '{}' to '{}' is {}years", date1, date2, y);
         LocalDateTime localDateTime1 = LocalDateTime.parse(date1, formatter);
         LocalDateTime localDateTime2 = LocalDateTime.parse(date2, formatter);
         assertEquals("Years differ", Long.parseLong(y.toString()), ChronoUnit.YEARS.between(localDateTime1, localDateTime2));
@@ -34,7 +35,7 @@ public class DateTimeSteps {
 
     @Then("DateTime check period from \"{}\" to \"{}\" is {}day(s)")
     public void compareDays(String date1, String date2, Object d) {
-        log.info("Check period from {} to {}", date1, date2);
+        logger.log("Check period from '{}' to '{}' is {}days", date1, date2, d);
         LocalDateTime localDateTime1 = LocalDateTime.parse(date1, formatter);
         LocalDateTime localDateTime2 = LocalDateTime.parse(date2, formatter);
         assertEquals("Days differ", Long.parseLong(d.toString()), ChronoUnit.DAYS.between(localDateTime1, localDateTime2));
@@ -42,7 +43,7 @@ public class DateTimeSteps {
 
     @Then("DateTime check period from \"{}\" to \"{}\" is {}hour(s)")
     public void compareHours(String date1, String date2, Object h) {
-        log.info("Check period from {} to {}", date1, date2);
+        logger.log("Check period from '{}' to '{}' is {}hours", date1, date2, h);
         LocalDateTime localDateTime1 = LocalDateTime.parse(date1, formatter);
         LocalDateTime localDateTime2 = LocalDateTime.parse(date2, formatter);
         assertEquals("Hours differ", Long.parseLong(h.toString()), Duration.between(localDateTime1, localDateTime2).toHours());
@@ -50,7 +51,7 @@ public class DateTimeSteps {
 
     @Then("DateTime check period from \"{}\" to \"{}\" is {}minute(s)")
     public void compareMinutes(String date1, String date2, Object m) {
-        log.info("Check period from {} to {}", date1, date2);
+        logger.log("Check period from '{}' to '{}' is {}mins", date1, date2, m);
         LocalDateTime localDateTime1 = LocalDateTime.parse(date1, formatter);
         LocalDateTime localDateTime2 = LocalDateTime.parse(date2, formatter);
         assertEquals("Minutes differ", Long.parseLong(m.toString()), Duration.between(localDateTime1, localDateTime2).toMinutes());
@@ -58,7 +59,7 @@ public class DateTimeSteps {
 
     @Then("DateTime check period from \"{}\" to \"{}\" is {}second(s)")
     public void compareSeconds(String date1, String date2, Object s) {
-        log.info("Check period from {} to {}", date1, date2);
+        logger.log("Check period from '{}' to '{}' is {}seconds", date1, date2, s);
         LocalDateTime localDateTime1 = LocalDateTime.parse(date1, formatter);
         LocalDateTime localDateTime2 = LocalDateTime.parse(date2, formatter);
         assertEquals("Seconds differ", Long.parseLong(s.toString()), Duration.between(localDateTime1, localDateTime2).getSeconds());
@@ -66,17 +67,17 @@ public class DateTimeSteps {
 
     @Then("Date check period from \"{}\" to \"{}\" is {}year(s)")
     public void compareYearsFromDates(String date1, String date2, Object y) {
-        log.info("Check period from {} to {}", date1, date2);
+        logger.log("Check date period from '{}' to '{}' is {}years", date1, date2, y);
         LocalDate localDate1 = LocalDateTime.parse(date1, formatter).toLocalDate();
         LocalDate localDate2 = LocalDateTime.parse(date2, formatter).toLocalDate();
         assertEquals("Years differ", Long.parseLong(y.toString()), ChronoUnit.YEARS.between(localDate1, localDate2));
     }
 
     @Then("Date check period from \"{}\" to \"{}\" is {}day(s)")
-    public void compareDaysFromDates(String date1, String date2, Object y) {
-        log.info("Check period from {} to {}", date1, date2);
+    public void compareDaysFromDates(String date1, String date2, Object d) {
+        logger.log("Check date period from '{}' to '{}' is {}days", date1, date2, d);
         LocalDate localDate1 = LocalDateTime.parse(date1, formatter).toLocalDate();
         LocalDate localDate2 = LocalDateTime.parse(date2, formatter).toLocalDate();
-        assertEquals("Days differ", Long.parseLong(y.toString()), ChronoUnit.DAYS.between(localDate1, localDate2));
+        assertEquals("Days differ", Long.parseLong(d.toString()), ChronoUnit.DAYS.between(localDate1, localDate2));
     }
 }