+++ src/main/java/com/cucumber/utils/context/stepdefs/shell/ShellSteps.java
@@ -2,6 +2,7 @@ package com.cucumber.utils.context.stepdefs.shell;
 
 import com.cucumber.utils.clients.shell.ShellClient;
 import com.cucumber.utils.context.utils.Cucumbers;
+import com.cucumber.utils.context.utils.ScenarioUtils;
 import com.google.inject.Inject;
 import io.cucumber.guice.ScenarioScoped;
 import io.cucumber.java.en.Then;
@@ -11,10 +12,13 @@ public class ShellSteps {
 
     @Inject
     private Cucumbers cucumbers;
+    @Inject
+    private ScenarioUtils logger;
     private ShellClient shellClient = new ShellClient();
 
     @Then("SHELL execute command \"{}\" and check response=\"{}\"")
     public void executeAndCompare(String cmd, String expected) {
+        logger.log("Execute cmd '{}' and compare response with: {}", cmd, expected);
         String actual = shellClient.command("bash", "-c", cmd).trim();
         cucumbers.compare(expected, actual);
     }