+++ src/main/java/com/cucumber/utils/context/utils/ScenarioUtils.java
@@ -1,0 +1,24 @@
+package com.cucumber.utils.context.utils;
+
+import io.cucumber.guice.ScenarioScoped;
+import io.cucumber.java.Before;
+import io.cucumber.java.Scenario;
+import org.apache.logging.log4j.message.ParameterizedMessage;
+
+@ScenarioScoped
+public class ScenarioUtils {
+    private Scenario scenario;
+
+    @Before
+    public void init(Scenario scenario) {
+        this.scenario = scenario;
+    }
+
+    public void log(String msg, Object... args) {
+        scenario.write(ParameterizedMessage.format(msg, args));
+    }
+
+    public Scenario getScenario() {
+        return scenario;
+    }
+}